fix(Navigation): Fix "Sessao Noivado" with external link
Change the component AnchorToPagesButton to accept link or href, and make a decisiion if it will be an external link or not. If it's an external link, use regular anchor with a target (for this case, other page), if not it will use vue-router for the case. This changes will be applied on Navigation, and on Hompage.